requirements.txt
asgiref==3.6.0
Django==4.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
PyJWT==2.6.0
pytz==2023.3
sqlparse==0.4.3
mysqlclient==2.1.1
drf-yasg==1.21.5
django-silk==5.0.3
django-rest-passwordreset==1.3.0
